fix: trigger event
